+++ docs/source/conf.py
@@ -25,8 +25,8 @@ _build_lang = os.environ.get("SPHINX_LANG", "en")
 project = "slimv"
 copyright = "2026, Haleem"
 author = "Haleem"
-release = "0.2.2"
-version = "0.2.2"
+release = "0.3.0"
+version = "0.3.0"
 
 language = _build_lang
 html_title = f"slimv {release}"